Addressed missing function prototype warnings.
In C, a function that accepts no arguments should have a (void)
parameter list in its prototype, not an empty () one.  While they are
equivalent in C++, they are not the same in C.  This patch set address
missing prototype related warnings in C code that uses the functions
in question.

Change-Id: Ic5bb65eefd857da03c39d135fd68e43880e7b975
Bug-Tizen: TC-1208
Signed-off-by: Ossama Othman <ossama.othman@intel.com>