ilmControl: control_thread uses eventfd for shutdown signal
Do not use pthread_cancel to shutdown control_thread. This could cause
stale locks in libwayland-client when the cancel was received when
the display lock (according to the current 1.3 implementation) is held.

To alleviate this, one could disable thread_cancellation in
control_thread during wl_* calls, this however is one more hack to the
cancel-to-shutdown hack. Instead, use a clean interface to send an
event to the control_thread when it's time to shutdown.

Signed-off-by: Marcus Fritzsch <marcus.fritzsch@xse.de>