ilmControl: add ilm_registerNotification and ilm_unregisterNotification
ilm_registerNotification API registers a callback to ilmControl,
which is called when a layer or a surface is created or destroyed.
A void pointer of user_data can be also registered with this API,
which helps the applications to classify the notification.

ilm_unregisterNotification API unregisters the above mentioned callback
and user data

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>