ilmControl: set original dimensions at configuration event.
It is wrong to set it at source_rectangle event. Original dimensions
are dependent on the wayland buffer of the surface, and it can be modified
from the client. The compositor sends a configured event when
the buffer is modified. Therefore, we should also modify the original dimensions
at configuration event.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>