layer-add-surface: initial implementation
it has 2 parameters:

layerID: The layer with the given ID is created and added to the screen 0.
Moreover, it is made visible

number_of_surfaces: number of surfaces which should be added to the layer.

when the surface creation callback is called, the surface is made visible,
added to the layer. Moreover, destination rectangle and source rectangle are set
to size of its wayland buffers.

layer-add-surfaces exits when given number of surfaces are added to the layer

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>