protocol objects: Avoid relocations in .text
The current build system is compiling the scanner-generated code as an
intermediate static library, and then linking it in with the output shared
libraries.
This results in some code that was not build with -fPIC, which prevents
hardened non-writable code and produces warnings in buildsystems such as
yocto.

This patch adds -fPIC to the ivi-extension-protocol intermediate library.

Signed-off-by: <frederico.cadete@awtce.be>