ilmCommon: ilm_destroy() to be thread-safe and introduce an initialization counter
If ilm_init() is called multiple times, the ilm context should not be destroyed
at the first call of ilm_destroy(). Because a thread can call ilm_destroy() and destroy
commonly used ilm context.

Each ilm_init() call increments the initialization counter, and each ilm_destroy() call
decrements the counter. The ilm context is only destroyed when the counter is 0.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>