tests: set new expected values for notification tests
After a callback function for a surface is registered with the ilm_surfaceAddNotificaiton API,
the callback function is called with ILM_NOTIFICATION_CONTENT_AVAILABLE flag, if the surface exists.
This causes that the notification test fail. Because they are only expecting the flag of the
changed parameter (e.g.ILM_NOTIFICATION_ORIENTATION).

The notification tests are successful again after these changes.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>