ilmControl: remove failed to get surface/layer context in ilmControl error message
When an ilmControl session gets the controller_listener_surface event
after an application created an ivi_surface, it controls the existence of the
surface with the get_surface_context internal API. It is expected
that the surface does not exist.

The same is true also for wayland_controller_get_layer_context API.
Therefore, these error messages are redundant and confusing.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>