ivi-controller: don't send wrong layer resource to client
It is wrong to send surface_send_layer event with resources
of every client. Instead, the event should only sent once with
the layer resource of client, which owns also the surface resource.
Otherwise a wayland communication error happens, because a client gets
a surface_send_layer event with a resource which is owned by
another client.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>