ilmClient: mark all APIs deprecated
Users can use ivi-application protocol directly instead
of using ilmClient APIs. The APIs do not offer any advantage
over the direct use. Furthermore, ilm_surfaceCreate API is
misleading the users, because it is not possible to set width,
height and pixelformat of a surface with this API.

Therefore, ilmClient APIs are marked as deprecated.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>