LayerManagerControl: remove surface create and destroy api's
api's never worked and would even cause exception in LayerManagerControl tool
surface can only be created with native content and native content is
wayland-surface which is only available in the client.
So it does not make any sense to provide this functionality to external tool.
For surface destroy it is very similar issue, we can only destroy surfaces
which were created in the same client, but this is never the case by using
LayerManagerControl- every call will create new connection and therefore new client

Signed-off-by: Eugen Friedrich <efriedrich@de.adit-jv.com>