ilmControl: update startup and destroy
We have now a listener for wl_global_remove events,
so that we can destroy screens when their wl_output is
destroyed.

We do not need to roundtrip 3 times, because compositor
does not send properties of layer/surface. Furthermore,
we do not need to listen surface/layer creation, because
we create one controller surface and one controller layer.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>