ivi-controller: update module startup and signals
We are listening three additional signals:
1. destroy_signal of the compositor:
   Free all dynamic allocated memory and exit
2. output_created_signal of the compositor:
   Create an ivi-screen
3. output_destroyed_signal of the compositor:
   Destroy the ivi-screen

Furthermore, it is possible to predefine screen-ids.
The ids are given in weston.ini. An example weston.ini:

[core]
shell=ivi-shell.so

[ivi-shell]
ivi-module=ivi-controller.so
ivi-input-module=ivi-input-controller.so
screen-id-offset=10

[ivi-screen]
screen-name=LVDS-1
screen-id=0

[ivi-screen]
screen-name=HDMI-A-1
screen-id=1

[ivi-screen]
screen-name=HDMI-A-2
screen-id=2

[ivi-screen]
screen-name=VGA-1
screen-id=3

If an output is not defined in weston-ini,
its id is generated. Generated ids start from
<screen-id-offset> and incremented for every output

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>