ilmControl: introduce ilm_getError API
Some requests to compositor can fail because
of reasons that we cannot control,
e.g.: application destroys the surface.
Therefore, compositor sends error events when
a request fails. If we try to get these errors
during Set APIs, we have to roundtrip wayland
connection to get current error message.
This will decrease the performance.

Instead we can have the API ilm_getError to request
for error when we want.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>