ilmControl: introduce ilm_surfaceSetType API
If a surface is restricted type, visible contents of the surface is strictly
controlled by the compositor. Its content is not allowed to be go out of
its destination region. If the application resizes its buffers or uses
wp_viewporter protocol to scale its contents, the old destination region
would causes visible glitches.
To avoid these issues, the controller process mark a surface as desktop
compatible. Source and destination regions of a desktop compatible
surface will be modified accordingly,when application sends a request
for resizing or scaling its contents. Therefore, applications contents
will be drawn according to application's wishes.
On the other hand, source and destination regions will be strictly
enforced, when the surface's type is restricted. The default type for
a surface is restricted.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>