protocol: change screenshot requests in ivi-wm protocol
The screenshot requests in the ivi-controller protocol require the
server to write a file in a certain format. The server needs to support
one or more image formats and a client which wants to read those images
needs to be able to read those formats. There are no further parameters
to influence the used image formats and there are no events informing
about the supported image formats. Furthermore there is no feedback to
the client on when the image is written.
This change intends to remove those uncertainties, and also remove the
burden from the compositor to support any image formats.
With the new requests in the ivi-wm protocol the client will receive raw
pixeldata through a filedescriptor together with width, height, stride
and a wl_shm.format parameter.

The filename parameters from request ivi_wm_screen.screenshot and
request ivi_wm.surface_screenshot request are removed. Those requests
create now a new object which will return the screenshot result to the
client. These objects are of type ivi_screenshot which is also
introduced in this commit. An ivi_screenshot shall receive exactly on
event, either a 'done' event with the screenshot data or an 'error'
event when the screenshot failed.

This commit adapts also the screenshot request calls in the ilm API and
the corresponding listener callbacks in the weston-ivi-shell but leaves
them in an unimplemented state for now.

Signed-off-by: Mathias Fiedler <mathias_fiedler@mentor.com>