weston-ivi-shell, ilmControl: move bitmap save function from ivi-shell to ilm API
With the change of the ivi-wm screenshot requests the weston-ivi-shell
no longer writes bmp files when screenshots are requested. To restore
the behavior of having bmp files written through the ilmControl API the
ilmControl implementation itself needs to write the bmp files.

This commit moves the save_as_bitmap function from the weston-ivi-shell
to the ilmControl.

Signed-off-by: Mathias Fiedler <mathias_fiedler@mentor.com>