weston-ivi-shell: implement new ivi-wm screenshot requests
With the new ivi-wm screenshot requests the server needs to send either
a 'done' or an 'error' event on the ivi_screenshot object which is
created with the screenshot request.

The done event contains a filedescriptor for a file containing raw
pixeldata. The event also has parameters describing width, height,
stride and format of the pixeldata.
In case of an error an 'error' event is send instead.

Two new local helper functions are introduced, create_screenshot_file()
and flip_y().
Function create_screenshot_file() will create a file with temporary name
in the XDG_RUNTIME_DIR. It will furthermore resize the file to the
required size using fallocate or ftruncate similar to
os_create_anonymous_file in weston.
Function flip_y flips the pixel data upside down. This helper is used
for  pixeldata read from outputs when needed.

Signed-off-by: Mathias Fiedler <mathias_fiedler@mentor.com>