ilmCommon: remove unused parts of ilm_types.h
ilmLayerType struct is removed because hardware
layers are not supported.

ilmOrientation is removed because orientation API is
removed.

Pixelformat, orientation and nativeSurface properties are
not applicable for surfaces. Therefore, they are removed from
ilmSurfaceProperties struct.

origSourceWidth/Height, orientation, type and creatorPid
properties are not applicable for layers. Therefore, they
are removed from ilmLayerProperties struct.

There are no Health, Communicator, SceneProvider or Renderer
plugin. All related structs and defines are removed.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>