ivi-input: Clear touch focus if surface does not accept events from the seat
For touch device libweston sets the focus on first touch down event.
If the particular focus surface does not accept events from the
seat then, clear the touch focus. It avoids forwarding of events
to this surface in future. Also this is the right status indicating
touch focus is NULL.

Signed-off-by: harsha.mm <harsha.manjulamallikarjun@in.bosch.com>