ivi-input: Evaluate pointer focus and clear if surface does not accept events from the seat
Evaluation pointer focus in the pointer focus grab interface.
If the focus surface does not accept events from the seat then,
clear the pointer focus. NULL pointer focus shows the right
state and also simplies handling of further pointer events.

It should be noted that if pointer focus is NULL, it is not
possible to set the pointer cursor. Whereas if there is already
an existing cursor then the cursor can hover on the surface. It is
just that the set of pointer cursor is not possible if focus is
NULL.

Signed-off-by: harsha.mm <harsha.manjulamallikarjun@in.bosch.com>