ivi-input: Fix multi-seat focus setting and change setup of device focus
Focus should be separate for each seat. The focus set then is
applicable for that particular seat only. The bug related to
this is fixed.
Also there is change in setting of pointer and touch focus
since the pointer focus evaluation happens based on motion now
and touch/pointer focuses are cleared if surface does not
accept input from particular seat.

Signed-off-by: harsha.mm <harsha.manjulamallikarjun@in.bosch.com>