simple-ivi-share: update share surface when client restarts
Currently, simple-ivi-share cannot update share surface in case that
the client such as simple-egl is killed and restarted. Therefore,
this case should be considered because it is possible some applications
actually could be killed and restarted by user or system problem.

Originally, simple-ivi-share does not get a share surface any more
when the client is killed, so memory leak does not occur. But, during
implementation this feature, I found that the memory leak could be happened
when getting a share surface if the original surface does not exist.

When the original surface does not exist, ivi-share creates an empty
nativesurface and sets the state of share surface as not exist.
This empty nativesurface is still exist in memory, so memory leak could be occured.
Therefore, ivi_share_surface_destroy() should be called explicitly.