ivi-share: implement release_shared_name
When name is sent to the client, lock the buffer using weston_buffer_rederence.
When receiving the release request from the client, release the buffer using
weston_buffer_rederence.

This implementation expects that if the client receives a new name it will
release a old name. It is like "swap". Therefore, ivi-share manages two buffers,
"current" and "back".
In order to support to multiple clients, keep the number of clients sent buffers
for each buffer.
In order to support both protocol version v1 and v2, check the protocol version
of the client.

For only v1 client, only the "current" is used. In this case, basically it is
same behavior as before. There is a slight difference to hold the "current"
buffer until a new buffer is notified from weston.
For both v1 and v2 clients, it counts only the number of v2 clients, releases
the buffer when the number reaches 0 and sends a new name to the clients. This
means that the sending timing to the v1 client will be the same as the v2
client.When name is sent to the client, lock the buffer using weston_buffer_rederence.
When receiving the release request from the client, release the buffer using
weston_buffer_rederence.

This implementation expects that if the client receives a new name it will
release a old name. It is like "swap". Therefore, ivi-share manages two buffer
references.
In order to support to multiple clients, keep the number of clients sent buffers
for each buffer.
In order to support both protocol version v1 and v2, check the protocol version
of the client.

For only v1 client, only one buffer reference is used. In this case, basically
it is same behavior as before. There is a slight difference to hold the one
buffer rederence until a new buffer is notified from weston.
For both v1 and v2 clients, it counts only the number of v2 clients, releases
the buffer when the number reaches 0 and sends a new name to the clients. This
means that the sending timing to the v1 client will be the same as the v2
client.

Signed-off-by: Kenji Hosokawa <khosokawa@de.adit-jv.com>