ivi-input: send surface relative entry cooordinates when pointer focus is forced
Provide surface relative coordinates when pointer focus is forced
to a particular surface instead of providing(0,0). Forcing pointer
focus to a surface is particulary a use case of rotary knobs which
give out pointer axis events only. But applications can end up setting
pointer focus in a simulation environment where the touch input
is unavailable and pointer is used instead.

Signed-off-by: Harsha M M <harsha.manjulamallikarjun@in.bosch.com>