ivi-id-agent: added ivi-id-agent
This is a reference implementation of an ivi-id-agent plugin. It
creates surface-ids for desktop-surfaces depending on the configuration
provided in "weston.ini". For more please refer to the reference
implementation in this commit.

The only public interface available is the "id_agent_module_init"
function. It is responsible for initialization of structs and reading
the configuration.

In the reference "weston.ini" two types of configurations can be found:

[desktop-app] is used to configure a particular application. Therefore
the desired surface-id must be provided. Moreover "app-title" is the
title that is provided by the desktop application (xdg-protocol).
"app-id" behaves accordingly. Although both parameters can be set, it
is not mandatory. Finally at least one has to be set.

[desktop-app-default] enables the id-agent to generate generic
surface-ids for unconfigured applications, e.g. for development. This
tag is optional. To generate the id an interval starting from
"default-surface-id" to "default-surface-id-max" is used. The id is
incremented until the interval is exceeded.

In the function "get_id" a surface-id is assigned by means of the
configuration. It can be adjusted, if another behavior is desired.
In this plugin the parameters, that are described above, are evaluated.

To use these patches please also apply the dedicated patches for
weston: https://github.com/mtey/weston/tree/xdg_support_ivi_id_agent
Since libweston-desktop is used to introduce desktop-surface support,
xdg-protocol is supported.

Signed-off-by: Michael Teyfel <mteyfel@de.adit-jv.com>