wayland-ivi-extension: Correct the dependency order
Currently, ilmControl depends on ilmCommon. However, it is
ilmCommon that should depend on ilmControl.

Functions ilmControl_init and ilmControl_destroy are defined in
ilmControl, but are getting called from ilmCommon. Hence, ilmCommon
won't work without having ilmControl on the system at runtime. If
not found, there will be a runtime error which could be avoided
by linking these libraries correctly at build time.

Moreover, there is only a build time dependency of include headers
of ilmCommon on ilmControl. So at runtime, even if ilmCommon is not
present on the system, ilmControl can function independently.

This change corrects the dependency order of ilmCommon and
ilmControl.

Signed-off-by: Ashutosh Agarwal <asagarwal@nvidia.com>
Reviewed-by: Miguel Vico Moya <mvicomoya@nvidia.com>