LayerManagerControl: Prevent conversion of filenames into lowercase
LayerManagerControl auto-convert all text expressions to lowercase which includes
the dump filenames that the user provides.
LayerManagerControl should use the filename as provided and should not modify it.
Added necessary changes in LayerMangerControl source code so that the text is not
converted to lowercase if the expression is a filename.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>