ivi-controller: Avoid unnecessary scaling of the view transformation matrix
The opaque region of a weston view is updated only if the alpha value is 1
and the transform matrix is of type WESTON_MATRIX_TRANSFORM_TRANSLATE.

Here opaque region is never updated, as we are performing scaling operations
to the view transform matrix(even when the scaling factor is 1), thus
changing the matrix type to WESTON_MATRIX_TRANSFORM_SCALE.
As a result, views are never opaque and won't be able to place it in
HW planes(forced to use renderer).

Perform scaling of the view transformation matrix only when the scaling
factor is non-zero.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>