ivi-input-controller: Stop using frame_signal void *data parameter to get weston_output
Instead of getting weston_output from the frame_signal argument 'void *data',
add weston_output in the private data struct of the users that are listening
to frame_signal. With this change we are able to pass previous_damage as the
data argument.

This is similar to weston commit:
dd0d0b0d ("screenshooter: stop using frame_signal void *data parameter to get weston_output")

Signed-off-by: Marek Vasut <marex@denx.de>