EGLWLInputEventExample: Repair array out of bounds access
The problem is in the `fprintf(..., left_ptrs[j])`, where if this code is
ever triggered, the variable `j=4` always, while the `left_ptrs` array only
has four entries instead of five, so the code would access one entry past
the array. In case this code is triggered, it does cause segmentation fault.

Signed-off-by: Marek Vasut <marex@denx.de>