ilmControl: Extend screenshot API with callback support
Extend the screenshot API such that it permits user to add their own
hooks into screenshot_done and screenshot_err callbacks. This way,
the user can obtain the FD with the screenshot and process it instead
of having the screenshot written into a file in /tmp directory. Make
the filename optional, so the user can provide only the callbacks and
skip writing the file into /tmp altogether.

The library ABI is unaffected, the new functionality is added via two
library functions, ilm_takeScreenshot5() and ilm_takeSurfaceScreenshot5(),
which are now internally called by their original counterparts with
callback hooks set to NULL.

Signed-off-by: Marek Vasut <marex@denx.de>