CMake: Specify generic library destination for installation
The libraries generated are now installed in lib${LIB_SUFFIX} (lib/lib32/lib64)
directory. This doesn't work with Debian as it expects libraries to be present
in lib/<multiarch-tuple>.

Set the library destination path to 'CMAKE_INSTALL_LIBDIR'.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>

[Khang: Weston installed its modules to dir_module_weston, then loading
modules when run time. It's defined dir_lib/weston. Weston doesn't
puclib the variable via pkgconfig. But we can get the dir_lib via
WESTON_LIBDIR, then we can install the ivi-modules to that directory]

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>