ivi-controller: new mechanism for screenshot
On weston 11, with commit [1], the weston_renderer struct
becomes private. The ivi-controller uses read_pixels from
weston_renderer struct for screenshot.

To continue to support the screenshot, we need to switch
to the new mechanism. Using the weston_screenshooter_shoot api
is the best solution. With this api, ilm client needs to prepare
a shm buffer, then send it to the ivi-controller.

This change will not be compatible with older ivi-wm protocol
versions.

[1] https://gitlab.freedesktop.org/wayland/weston/-/commit/0c65b2384

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>