ilm_input_test: update tests with default seat
Currently, tests are using the 'default' as the default seat name.
In the case, the default seat name is not the 'default', tests will
return a failure.

ilm_getDefaultSeat is used to get the default seat name.
If it returned ILM_SUCCESS, execute the tests with
the seat name got from the API. Otherwise, skip the tests.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>